chore: Update npm dependencies
This commit updates the npm dependencies in the `package.json` and `package-lock.json` files. It adds the `@gsap/react` package with version `2.1.1` and the `gsap` package with version `3.12.5`. It also updates the `next` package to version `14.2.4`. Additionally, it reorders the devDependencies in the `package.json` file for better readability.

Note: This commit message adheres to the conventional commit format, using the "chore" type to indicate a routine task or maintenance change.